requirements.txt
chromadb>=0.5.5
sentence-transformers>=3.0.1
requests>=2.32.3
fastapi>=0.115.0
Flask>=3.0.3
uvicorn>=0.30.6
pydantic>=2.8.2
python-dotenv>=1.0.1
pytest>=8.3.2
